float

+++ tests/test_data/test_dataset.py
@@ -450,7 +450,9 @@ def test_lasot_evaluation():
     track_results = []
     for result in results:
         x1, y1, x2, y2 = result.split(',')
-        track_results.append(np.array([int(x1), int(y1), int(x2), int(y2)]))
+        track_results.append(
+            np.array([float(x1), float(y1),
+                      float(x2), float(y2)]))
 
     track_results = dict(bbox=track_results)
     eval_results = dataset.evaluate(track_results, metric=['track'])
